Do not invoke setupy.py directly.
invoking setup.py directly is not the best way to install a package from
source. See one of the most recent blog about this:

https://blog.ganssle.io/articles/2021/10/setup-py-deprecated.html

This updates the readme to invoke pip, and use `python -m pip` to make
sure that the pip which is invoked is the one from the current Python.

i'm also happy to make it `python3 -m pip`, but these days python points
to python 3 most of the time.